Change url param from domain_name to key
The change was made because as per the old logic, if a page within
the domain was submited, the url generated would be the same.

As per the new logic the url generated depends on the url submited and date. It
would help us to generate a unique URL, at the same time we can use this key
to build a cache system